requirements.txt
termcolor
jsonpickle
gitpython
tabulate
mdutils
blessings
requests